pyproject.toml
[project]
name = "indibyte"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "chandra-ocr>=0.1.7",
    "dspy-ai>=3.0.4",
    "fastapi>=0.121.2",
    "fastmcp>=2.13.0.2",
    "httpx>=0.28.1",
    "langchain>=1.0.5",
    "openai>=2.8.0",
    "openai-whisper>=20250625",
    "opencv-python-headless>=4.12.0.88",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "python-socketio>=5.14.3",
    "ruff>=0.14.4",
    "transformers>=4.57.1",
    "uvloop>=0.22.1",
    "websockets>=15.0.1",
]
